requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
Pillow==9.1.0
python-slugify==6.1.2
sqlparse==0.4.2
text-unidecode==1.3
